restructure code, solve broadcast 3a challenge
