+++ broadcast/go.mod
@@ -1,0 +1,5 @@
+module broadcast
+
+go 1.20
+
+require github.com/jepsen-io/maelstrom/demo/go v0.0.0-20230516124010-52951329816e
